Describe a gene-name text-search index

Builds an aggregate Trix text-search adapter so users can search the browser by gene name. Pass the result as text_search to JBrowseR().

text_index(ix, ixx, meta, assembly)

Arguments

ix

URL to the .ix file.

ixx

URL to the .ixx file.

meta

URL to the meta.json file.

assembly

Name of the assembly the index applies to.

Value

an aggregate text-search adapter list

Details

Hub assemblies (e.g. JBrowseR("hg38")) already include gene search, so this is only needed for custom assemblies with your own Trix index.

Examples

text_index(
  "https://jbrowse.org/genomes/hg19/trix/hg19.ix",
  "https://jbrowse.org/genomes/hg19/trix/hg19.ixx",
  "https://jbrowse.org/genomes/hg19/trix/meta.json",
  "hg19"
)
#> $type
#> [1] "TrixTextSearchAdapter"
#> 
#> $textSearchAdapterId
#> [1] "hg19-index"
#> 
#> $ixFilePath
#> $ixFilePath$uri
#> [1] "https://jbrowse.org/genomes/hg19/trix/hg19.ix"
#> 
#> 
#> $ixxFilePath
#> $ixxFilePath$uri
#> [1] "https://jbrowse.org/genomes/hg19/trix/hg19.ixx"
#> 
#> 
#> $metaFilePath
#> $metaFilePath$uri
#> [1] "https://jbrowse.org/genomes/hg19/trix/meta.json"
#> 
#> 
#> $assemblyNames
#> $assemblyNames[[1]]
#> [1] "hg19"
#> 
#>